Streamline version suggestions.
