+++ depscan/cli.py
@@ -27,7 +27,7 @@ from depscan.lib.analysis import (
     analyse_pkg_risks,
     find_purl_usages,
     prepare_vdr,
-    summary_stats, suggest_version,
+    summary_stats,
 )
 from depscan.lib.audit import audit, risk_audit, risk_audit_map, type_audit_map
 from depscan.lib.bom import (
@@ -311,7 +311,8 @@ def build_args():
     return parser.parse_args()
 
 
-def scan(project_type, pkg_list, suggest_mode):
+# Deprecated
+def scan(project_type, pkg_list):
     """
     Method to search packages in our vulnerability database
 
@@ -328,91 +329,8 @@ def scan(project_type, pkg_list, suggest_mode):
         LOG.debug("Empty package search attempted!")
     else:
         LOG.debug("Scanning %d oss dependencies for issues", len(pkg_list))
-    # if not suggest_mode and bom_file:
-    #     results = (search_by_cdx_bom(bom_file))
-    # pkg_aliases is a dict that can be used to find the original vendor and
-    # package name This way we consistently use the same names used by the
-    # caller irrespective of how the result was obtained
-    sug_version_dict = {}
     results, pkg_aliases, purl_aliases = utils.search_pkgs(project_type, pkg_list)
-    if suggest_mode:
-        # From the results identify optimal max version
-        sug_version_dict = suggest_version(results, pkg_aliases, purl_aliases)
-        if sug_version_dict:
-            LOG.debug(
-                "Adjusting fix version based on the initial suggestion %s",
-                sug_version_dict,
-            )
-            # Recheck packages
-            sug_pkg_list = []
-            for k, v in sug_version_dict.items():
-                if not v:
-                    continue
-                sug, aliases = process_suggestions(k, v)
-                if sug:
-                    sug_pkg_list.extend(sug)
-                if aliases:
-                    pkg_aliases |= aliases
-            LOG.debug(
-                "Re-checking our suggestion to ensure there are no further "
-                "vulnerabilities"
-            )
-            override_results, _, _ = utils.search_pkgs(project_type, sug_pkg_list)
-            if override_results:
-                new_sug_dict = suggest_version(override_results)
-                LOG.debug("Received override results: %s", new_sug_dict)
-                for nk, nv in new_sug_dict.items():
-                    sug_version_dict[nk] = nv
-    return results, pkg_aliases, sug_version_dict, purl_aliases
-
-
-def process_suggestions(k, v):
-    """
-    Processes suggestions for package information and returns a list of packages
-    along with their aliases.
-
-    :param k: Package URL
-    :param v: Suggested version
-
-    :returns: A list of packages and a dict of aliases
-    :rtype: tuple[list, dict]
-    """
-    vendor = ""
-    version = v
-    pkg_list = []
-    aliases = {}
-    # Key is already a purl
-    if k.startswith("pkg:"):
-        with contextlib.suppress(Exception):
-            purl_obj = parse_purl(k)
-            vendor = purl_obj.get("namespace", purl_obj.get("type"))
-            name = purl_obj.get("name")
-            version = purl_obj.get("version")
-            pkg_list.append(
-                {
-                    "vendor": vendor,
-                    "name": name,
-                    "version": version,
-                    "purl": k,
-                }
-            )
-    else:
-        tmp_a = k.split(":")
-        if len(tmp_a) == 3:
-            vendor = tmp_a[0]
-            name = tmp_a[1]
-        else:
-            name = tmp_a[0]
-        # De-alias the vendor and package name
-        full_pkg = f"{vendor}:{name}:{version}"
-        full_pkg = aliases.get(full_pkg, full_pkg)
-        split_pkg = full_pkg.split(":")
-        if len(split_pkg) == 3:
-            vendor, name, version = split_pkg
-        elif split_pkg:
-            name = split_pkg[0]
-        pkg_list.append({"vendor": vendor, "name": name, "version": version})
-    return pkg_list, aliases
+    return results, pkg_aliases, purl_aliases
 
 
 def summarise(
@@ -458,8 +376,6 @@ def summarise(
     )
     pkg_vulnerabilities, pkg_group_rows = prepare_vdr(options)
     vdr_file = bom_file.replace(".json", ".vdr.json") if bom_file else None
-    # TODO: add option that consolidates, must move analysis items to properties, alter those and existing properties to include relevant bom-ref
-    # pkg_vulnerabilities = consolidate_vdrs(pkg_vulnerabilities)
     if pkg_vulnerabilities and bom_file:
         try:
             with open(bom_file, encoding="utf-8") as fp:
@@ -599,7 +515,7 @@ async def run_scan():
     db = db_lib.get()
     profile = "generic"
     deep = False
-    suggest_mode = q.get("suggest")
+    suggest_mode = q.get("suggest") or True
     if q.get("url"):
         url = q.get("url")
     if q.get("path"):
@@ -719,12 +635,12 @@ async def run_scan():
             audit_results = audit(project_type, pkg_list)
             if audit_results:
                 results = results + audit_results
-        vdb_results, pkg_aliases, sug_version_dict, purl_aliases = scan(
-            project_type,
-            pkg_list,
-            True
-        )
-        results += vdb_results
+        if not pkg_list:
+            LOG.debug("Empty package search attempted!")
+        else:
+            LOG.debug("Scanning %d oss dependencies for issues", len(pkg_list))
+        vdb_results, pkg_aliases, purl_aliases = utils.search_pkgs(project_type, pkg_list)
+        results.extend(vdb_results)
         with open(bom_file_path, encoding="utf-8") as fp:
             bom_data = json.load(fp)
         if not bom_data:
@@ -1083,11 +999,11 @@ def main():
                     src_dir,
                     project_type,
                 )
-        vdb_results, pkg_aliases, sug_version_dict, purl_aliases = scan(
-            project_type,
-            pkg_list,
-            args.suggest,
-        )
+        if not pkg_list:
+            LOG.debug("Empty package search attempted!")
+        else:
+            LOG.debug("Scanning %d oss dependencies for issues", len(pkg_list))
+        vdb_results, pkg_aliases, purl_aliases = utils.search_pkgs(project_type, pkg_list)
         results.extend(vdb_results)
         direct_purls, reached_purls = find_purl_usages(
             bom_file, src_dir, args.reachables_slices_file
